Prevent crash from Billion Laughs attack.

+++ src/SixLabors.Fonts/Tables/AdvancedTypographic/GPos/LookupType1SubTable.cs
@@ -70,7 +70,7 @@ namespace SixLabors.Fonts.Tables.AdvancedTypographic.GPos
             GPosTable table,
             GlyphPositioningCollection collection,
             Tag feature,
-            ushort index,
+            int index,
             int count)
         {
             ushort glyphId = collection[index];
@@ -141,7 +141,7 @@ namespace SixLabors.Fonts.Tables.AdvancedTypographic.GPos
             GPosTable table,
             GlyphPositioningCollection collection,
             Tag feature,
-            ushort index,
+            int index,
             int count)
         {
             ushort glyphId = collection[index];
