+++ src/SixLabors.Fonts/Tables/AdvancedTypographic/GPos/LookupType3SubTable.cs
@@ -77,7 +77,7 @@ namespace SixLabors.Fonts.Tables.AdvancedTypographic.GPos
                 GPosTable table,
                 GlyphPositioningCollection collection,
                 Tag feature,
-                ushort index,
+                int index,
                 int count)
             {
                 if (count <= 1)
@@ -93,7 +93,7 @@ namespace SixLabors.Fonts.Tables.AdvancedTypographic.GPos
                     return false;
                 }
 
-                ushort nextIndex = (ushort)(index + 1);
+                int nextIndex = index + 1;
                 ushort nextGlyphId = collection[nextIndex];
                 if (nextGlyphId == 0)
                 {
