+++ src/SixLabors.Fonts/Tables/AdvancedTypographic/GSub/LookupType3SubTable.cs
@@ -87,7 +87,7 @@ namespace SixLabors.Fonts.Tables.AdvancedTypographic.GSub
             GSubTable table,
             GlyphSubstitutionCollection collection,
             Tag feature,
-            ushort index,
+            int index,
             int count)
         {
             ushort glyphId = collection[index];
